+++ lib/util/talk/talk_dialog.dart
@@ -10,6 +10,7 @@ class TalkDialog extends StatefulWidget {
     required this.says,
     this.finish,
     this.onChangeTalk,
+    this.backgroundColor,
     this.textBoxMinHeight = 100,
     this.keyboardKeyToNext,
     this.padding,
@@ -20,17 +21,20 @@ class TalkDialog extends StatefulWidget {
     List<Say> sayList, {
     VoidCallback? finish,
     ValueChanged<int>? onChangeTalk,
+    Color? backgroundColor,
     double boxTextHeight = 100,
     LogicalKeyboardKey? logicalKeyboardKeyToNext,
     EdgeInsetsGeometry? padding,
   }) {
     showDialog(
+      barrierColor: backgroundColor,
       context: context,
       builder: (BuildContext context) {
         return TalkDialog(
           says: sayList,
           finish: finish,
           onChangeTalk: onChangeTalk,
+          backgroundColor: backgroundColor,
           textBoxMinHeight: boxTextHeight,
           keyboardKeyToNext: logicalKeyboardKeyToNext,
           padding: padding,
@@ -42,6 +46,7 @@ class TalkDialog extends StatefulWidget {
   final List<Say> says;
   final VoidCallback? finish;
   final ValueChanged<int>? onChangeTalk;
+  final Color? backgroundColor;
   final double? textBoxMinHeight;
   final LogicalKeyboardKey? keyboardKeyToNext;
   final EdgeInsetsGeometry? padding;